requirement.txt
matplotlib                3.1.1
numpy                     1.19.1
opencv                    3.4.2
pillow                    7.2.0
python                    3.7.9
pytorch                   1.2.0

